# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SessãoInk API",
    "description": "API de gestão para tatuadores e estúdios de tatuagem",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/agenda/": {
      "get": {
        "tags": [
          "agenda"
        ],
        "summary": "Listar Agenda",
        "operationId": "listar_agenda_api_v1_agenda__get",
        "parameters": [
          {
            "name": "ano",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 2035,
              "minimum": 2020,
              "title": "Ano"
            }
          },
          {
            "name": "mes",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 12,
              "minimum": 1,
              "title": "Mes"
            }
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgendaMesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agenda/agendar": {
      "post": {
        "tags": [
          "agenda"
        ],
        "summary": "Agendar Sessao",
        "operationId": "agendar_sessao_api_v1_agenda_agendar_post",
        "parameters": [
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgendarSessaoRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessaoResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agenda/proximas": {
      "get": {
        "tags": [
          "agenda"
        ],
        "summary": "Proximas Sessoes",
        "operationId": "proximas_sessoes_api_v1_agenda_proximas_get",
        "parameters": [
          {
            "name": "dias",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 30,
              "minimum": 1,
              "default": 7,
              "title": "Dias"
            }
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SessaoResponse"
                  },
                  "title": "Response Proximas Sessoes Api V1 Agenda Proximas Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgendaMesResponse": {
        "properties": {
          "ano": {
            "type": "integer",
            "title": "Ano"
          },
          "mes": {
            "type": "integer",
            "title": "Mes"
          },
          "total_sessoes": {
            "type": "integer",
            "title": "Total Sessoes"
          },
          "sessoes": {
            "items": {
              "$ref": "#/components/schemas/SessaoResponse"
            },
            "type": "array",
            "title": "Sessoes"
          },
          "dias_com_sessao": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Dias Com Sessao"
          }
        },
        "type": "object",
        "required": [
          "ano",
          "mes",
          "total_sessoes",
          "sessoes",
          "dias_com_sessao"
        ],
        "title": "AgendaMesResponse"
      },
      "AgendarSessaoRequest": {
        "properties": {
          "atendimento_id": {
            "type": "string",
            "format": "uuid",
            "title": "Atendimento Id"
          },
          "data_sessao": {
            "type": "string",
            "format": "date-time",
            "title": "Data Sessao"
          },
          "duracao_minutos": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duracao Minutos",
            "default": 120
          }
        },
        "type": "object",
        "required": [
          "atendimento_id",
          "data_sessao"
        ],
        "title": "AgendarSessaoRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SessaoResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "data_sessao": {
            "type": "string",
            "format": "date-time",
            "title": "Data Sessao"
          },
          "duracao_minutos": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duracao Minutos"
          },
          "tipo": {
            "type": "string",
            "title": "Tipo"
          },
          "status_operacional": {
            "type": "string",
            "title": "Status Operacional"
          },
          "descricao": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descricao"
          },
          "estilo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estilo"
          },
          "parte_corpo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parte Corpo"
          },
          "valor_total": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valor Total"
          },
          "cliente_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cliente Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "data_sessao",
          "duracao_minutos",
          "tipo",
          "status_operacional",
          "descricao",
          "estilo",
          "parte_corpo",
          "valor_total",
          "cliente_id"
        ],
        "title": "SessaoResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}